[BE] feat: 지도 필터링 조건별 매장 목록 조회 API 구현

See merge request s11-fintech-finance-sub1/S11P21B302!54

+++ wannaeat-be/src/main/java/com/waterdragon/wannaeat/domain/reservation/domain/ReservationTable.java
@@ -14,6 +14,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Entity
 @Getter
@@ -33,7 +34,6 @@ public class ReservationTable {
 	private Reservation reservation;
 
 	@Column(name = "table_id", nullable = false)
-	private Long tableId;
+	private int tableId;
 
-	// 생성자, getter, setter
 }
